Created a utility for primitive types.
The utility exposes objects corresponding to primitive types and methods on those objects providing functionality similar to the prototype methods of their primitive, except they do not operate on the prototype but instead take a primitive to run the method on.  This allows these methods to be passed to iterator functions such as .map() in a succinct and readable manner.